project/requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-jazzmin==2.5.0
django-phonenumber-field==7.0.0
django-telegrambot==1.0.1
et-xmlfile==1.1.0
idna==3.4
openpyxl==3.0.10
phonenumbers==8.12.55
Pillow==9.2.0
pyTelegramBotAPI==4.7.0
python-telegram-bot==13.14
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
